requirements.txt
Django==1.6.2

## Socket
gevent==1.0
gevent-socketio==0.3.6
gevent-websocket==0.9.3
